defined a new type samplingMechanism

+++ pkg/trace/sampler/dynamic_config.go
@@ -46,7 +46,9 @@ type rmc struct {
 // one can read/write on it concurrently, using getters and setters.
 type RateByService struct {
 	mu sync.RWMutex // guards rates
-	// currentColor is either 0 or 1.
+	// currentColor is either 0 or 1. And, it changes every time `SetAll()` is called.
+	// When `SetAll()` is called, we paint affected keys with `currentColor`.
+	// If there is a key has a color doesn't match `currentColor`, it means that key no longer exists.
 	currentColor int8
 	rates        map[string]*rmc
 	version      string
@@ -103,7 +105,7 @@ func (rbs *RateByService) GetNewState(version string) State {
 	for k, v := range rbs.rates {
 		ret.Rates[k] = v.r
 		if v.m != 0 {
-			ret.Mechanisms[k] = v.m
+			ret.Mechanisms[k] = uint32(v.m)
 		}
 	}
